# Makes IonRawBinaryWriter reuse ContainerInfo objects, reducing allocations and garbage.