*Description of changes:*

`ContainerInfo` is a mutable internal class that is used to maintain state about a container, annotation, decimal, or timestamp. Instances are stored in a stack, previously using a `LinkedList`. A new instance was created and pushed onto the stack on every `stepIn`, `writeDecimal` (except when the value was null or zero), `writeTimestamp` (except when the value was null), and whenever a value had at least one annotation. On each corresponding `stepOut`, or on the completion of the value, the instance at the top of the stack was popped and released for garbage collection. This led to `ContainerInfo`s and the `LinkedList.Node`s used to wrap them accounting for a significant portion of the total allocations in applications that write long streams of binary Ion data.

Real-world Ion data is typically fairly shallow (container depth < 8), and most Ion streams are fairly homogeneous in terms of structure (i.e. the container depth is not going to vary widely between top-level values). This makes the stack a good candidate for representation using an `ArrayList` with an initial size of 10. In many applications, it will never need to grow. This eliminates the `LinkedList.Node` wrappers.

Additionally, `ContainerInfo` objects are a good candidate for reuse, as they are already mutable and do not need to be thread-safe. Instead of dropping the reference to a `ContainerInfo` instance when popping from the stack, the stack pointer will simply be decremented and the instance left alive. The next time the stack grows to that depth, the instance will be reinitialized with new data. This eliminates the continual allocation and garbage collection of `ContainerInfo` objects.

In a benchmark, this resulted in a slight speedup (5%) and reduction in heap size (4%), and more significant reductions in GC count (56%), GC time (13%), and Eden Space churn (71%). Because more `ContainerInfo` objects are surviving GC, Survivor Space increased slightly from 0.055MB/s to 0.097MB/s. Full results below.

```
Old

Benchmark                                                                Mode  Cnt          Score          Error   Units
IonJavaContainersBenchmark.containers                                    avgt   20      14030.209 ±      205.426   ms/op
IonJavaContainersBenchmark.containers:Heap usage                         avgt   20       1275.527 ±        0.365      MB
IonJavaContainersBenchmark.containers:·gc.churn.{PS Eden Space}          avgt   20        709.896 ±       10.401  MB/sec
IonJavaContainersBenchmark.containers:·gc.churn.{PS Eden Space}.norm     avgt   20  744380153.163 ± 10906741.367    B/op
IonJavaContainersBenchmark.containers:·gc.churn.{PS Old Gen}             avgt   20            ≈ 0                 MB/sec
IonJavaContainersBenchmark.containers:·gc.churn.{PS Old Gen}.norm        avgt   20            ≈ 0                   B/op
IonJavaContainersBenchmark.containers:·gc.churn.{PS Survivor Space}      avgt   20          0.055 ±        0.042  MB/sec
IonJavaContainersBenchmark.containers:·gc.churn.{PS Survivor Space}.norm avgt   20      57437.375 ±    44536.127    B/op
IonJavaContainersBenchmark.containers:·gc.count                          avgt   20        180.000                 counts
IonJavaContainersBenchmark.containers:·gc.time                           avgt   20      88291.000                     ms

New

Benchmark                                                                Mode  Cnt          Score          Error   Units
IonJavaContainersBenchmark.containers                                    avgt   20      13310.065 ±      103.225   ms/op
IonJavaContainersBenchmark.containers:Heap usage                         avgt   20       1224.382 ±       10.605      MB
IonJavaContainersBenchmark.containers:·gc.churn.{PS Eden Space}          avgt   20        208.660 ±        1.929  MB/sec
IonJavaContainersBenchmark.containers:·gc.churn.{PS Eden Space}.norm     avgt   20  218795380.000 ±  2022982.275    B/op
IonJavaContainersBenchmark.containers:·gc.churn.{PS Old Gen}             avgt   20            ≈ 0                 MB/sec
IonJavaContainersBenchmark.containers:·gc.churn.{PS Old Gen}.norm        avgt   20            ≈ 0                   B/op
IonJavaContainersBenchmark.containers:·gc.churn.{PS Survivor Space}      avgt   20          0.097 ±        0.036  MB/sec
IonJavaContainersBenchmark.containers:·gc.churn.{PS Survivor Space}.norm avgt   20     101684.014 ±    37547.912    B/op
IonJavaContainersBenchmark.containers:·gc.count                          avgt   20         80.000                 counts
IonJavaContainersBenchmark.containers:·gc.time                           avgt   20      76442.000                     ms
``` 

These results were generated by the following JMH benchmark code:

```java
import com.amazon.ion.IonType;
import com.amazon.ion.IonWriter;
import com.amazon.ion.Timestamp;
import com.amazon.ion.system.IonBinaryWriterBuilder;
import org.openjdk.jmh.annotations.Benchmark;
import org.openjdk.jmh.annotations.Level;
import org.openjdk.jmh.annotations.Mode;
import org.openjdk.jmh.annotations.Scope;
import org.openjdk.jmh.annotations.Setup;
import org.openjdk.jmh.annotations.State;
import org.openjdk.jmh.annotations.TearDown;
import org.openjdk.jmh.infra.BenchmarkParams;
import org.openjdk.jmh.infra.IterationParams;
import org.openjdk.jmh.profile.GCProfiler;
import org.openjdk.jmh.profile.InternalProfiler;
import org.openjdk.jmh.profile.ProfilerResult;
import org.openjdk.jmh.results.AggregationPolicy;
import org.openjdk.jmh.results.IterationResult;
import org.openjdk.jmh.results.Result;
import org.openjdk.jmh.runner.Runner;
import org.openjdk.jmh.runner.options.OptionsBuilder;

import java.io.IOException;
import java.io.OutputStream;
import java.lang.management.ManagementFactory;
import java.math.BigDecimal;
import java.util.ArrayList;
import java.util.Collection;
import java.util.List;
import java.util.concurrent.TimeUnit;

@State(Scope.Benchmark)
public class IonJavaContainersBenchmark {

    private IonWriter writer;
    private int bytesWritten = 0;

    @Setup(Level.Iteration)
    public void setup() {
        writer = IonBinaryWriterBuilder.standard().build(new OutputStream() {
            @Override
            public void write(int b) {
                bytesWritten++;
            }
        });
    }

    @TearDown(Level.Iteration)
    public void tearDown() throws IOException {
        writer.close();
        bytesWritten = 0;
    }

    @Benchmark
    public int containers() throws Exception {
        for (int i = 0; i < 10000000; i++) {
            writer.stepIn(IonType.STRUCT);
            writer.addTypeAnnotation("annotation1");
            writer.setFieldName("list");
            writer.stepIn(IonType.LIST);
            writer.writeTimestamp(Timestamp.valueOf("2000T"));
            writer.stepIn(IonType.LIST);
            writer.stepIn(IonType.LIST);
            writer.stepIn(IonType.LIST);
            writer.stepIn(IonType.LIST);
            writer.stepIn(IonType.LIST);
            writer.stepOut();
            writer.stepOut();
            writer.stepOut();
            writer.stepOut();
            writer.stepOut();
            writer.stepOut();
            writer.setFieldName("sexp");
            writer.stepIn(IonType.SEXP);
            writer.stepIn(IonType.STRUCT);
            writer.setFieldName("decimal");
            writer.writeDecimal(BigDecimal.valueOf(3.14159));
            writer.stepOut();
            writer.stepOut();
            writer.stepOut();
        }
        writer.finish();
        return bytesWritten;
    }

    public static class HeapProfiler implements InternalProfiler {

        @Override
        public boolean checkSupport(List<String> list) {
            return true;
        }

        @Override
        public String label() {
            return "HEAP_USAGE";
        }

        @Override
        public String getDescription() {
            return "Heap usage profiler";
        }

        @Override
        public void beforeIteration(BenchmarkParams benchmarkParams, IterationParams iterationParams) {
            // Nothing.
        }

        @Override
        public Collection<? extends Result> afterIteration(BenchmarkParams benchmarkParams, IterationParams iterationParams,
                                                           IterationResult result) {

            long totalHeap = ManagementFactory.getMemoryMXBean().getHeapMemoryUsage().getUsed();

            Collection<Result> results = new ArrayList<>();
            results.add(new ProfilerResult("Heap usage", totalHeap / 1e6, "MB", AggregationPolicy.AVG));

            return results;
        }
    }

    public static void main(String[] args) throws Exception {
        new Runner(
            new OptionsBuilder()
                .include(IonJavaContainersBenchmark.class.getSimpleName())
                .addProfiler(HeapProfiler.class)
                .addProfiler(GCProfiler.class)
                .mode(Mode.AverageTime)
                .measurementIterations(10)
                .warmupIterations(10)
                .forks(2)
                .shouldDoGC(true)
                .timeUnit(TimeUnit.MILLISECONDS)
                .build()
        ).run();
    }
}
```

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.