@tgregg 2019-05-31
> This looks like it would be useful in other places as well. You should consider extracting it into its own generic class so the implementation details don't have to live in `IonRawBinaryWriter`. That would also make it easier to add a size ceiling should we ever decide to.

See #245